Reorg the classes
