+++ brainwallet/Export Classes/ExportButtonViewModel.swift
